+++ lib/src/main/java/com/smileidentity/results/SmileIDResult.kt
@@ -77,7 +77,7 @@ data class EnhancedKycResult(
  */
 @Parcelize
 data class DocumentVerificationResult(
-    val selfieFile: File? = null, // TODO Unmark as optional when selfie capture will be implemented
+    val selfieFile: File,
     val documentFrontFile: File,
     val documentBackFile: File? = null,
     val jobStatusResponse: DocVJobStatusResponse,