+++ lib/src/main/java/com/smileidentity/viewmodel/DocumentViewModel.kt
@@ -102,7 +102,7 @@ class DocumentViewModel(
                 when (result) {
                     is ImageCaptureResult.Error -> it.resumeWithException(result.throwable)
                     is ImageCaptureResult.Success -> it.resume(
-                        postProcessImage(file = documentFile),
+                        postProcessImage(file = documentFile, desiredAspectRatio = idAspectRatio),
                     )
                 }
             }
@@ -153,7 +153,9 @@ class DocumentViewModel(
             val prepUploadResponse = SmileID.api.prepUpload(prepUploadRequest)
             val frontImageInfo = documentFrontFile.asDocumentImage()
             val backImageInfo = documentBackFile?.asDocumentImage()
-            val selfieImageInfo = selfieFile?.asSelfieImage()
+            val selfieImageInfo = selfieFile?.asSelfieImage() ?: throw IllegalStateException(
+                "Selfie file is null",
+            )
             val uploadRequest = UploadRequest(
                 images = listOfNotNull(frontImageInfo, backImageInfo, selfieImageInfo),
             )
@@ -171,6 +173,7 @@ class DocumentViewModel(
             val jobStatusResponse = SmileID.api.getDocVJobStatus(jobStatusRequest)
             result = SmileIDResult.Success(
                 DocumentVerificationResult(
+                    selfieFile = selfieImageInfo.image,
                     documentFrontFile = documentFrontFile,
                     documentBackFile = documentBackFile,
                     jobStatusResponse = jobStatusResponse,