+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Made IdInfo Parcelable
 - Add helper functions which return a Flow of the latest JobStatus for a Job until it is complete
 - Add a `JobStatusResponse` interface
+- Option to disable Instructions Screen on Document Verification and SmartSelfie™
 
 ### Fixed